feat: criando os services das entidades e estruturando as entidades para o banco de dados

+++ src/main/java/com/DesafioSparta/Eduardo/dto/TaxaResponse.java
@@ -3,6 +3,6 @@ package com.DesafioSparta.Eduardo.dto;
 import java.math.BigDecimal;
 
 public record TaxaResponse(
-    BigDecimal percentual
+    double percentual
 ) {
 }
